+++ src/components/SplashScreen.tsx
@@ -1,0 +1,46 @@
+import React, { useEffect, useState } from 'react';
+import { motion, AnimatePresence } from 'framer-motion';
+
+interface SplashScreenProps {
+  onAnimationComplete: () => void;
+}
+
+const SplashScreen: React.FC<SplashScreenProps> = ({ onAnimationComplete }) => {
+  const [isVisible, setIsVisible] = useState(true);
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setIsVisible(false);
+      setTimeout(onAnimationComplete, 1000); // Wait for fade out animation to complete
+    }, 2000); // Show splash screen for 2 seconds
+
+    return () => clearTimeout(timer);
+  }, [onAnimationComplete]);
+
+  return (
+    <AnimatePresence>
+      {isVisible && (
+        <motion.div
+          initial={{ opacity: 1 }}
+          exit={{ opacity: 0 }}
+          transition={{ duration: 1 }}
+          className="fixed inset-0 z-50 flex items-center justify-center bg-blackboard-black"
+        >
+          <motion.div
+            initial={{ opacity: 0, scale: 0.8, filter: "blur(10px)" }}
+            animate={{ opacity: 1, scale: 1, filter: "blur(0px)" }}
+            transition={{ duration: 0.8 }}
+            className="text-white text-4xl md:text-6xl font-light"
+          >
+            <div className="flex flex-col items-start justify-center text-sm md:text-md">
+              <div className="text-slate font-light">Work</div>
+              <div className="text-white  font-light">by Alen Koikkara.</div>
+            </div>
+          </motion.div>
+        </motion.div>
+      )}
+    </AnimatePresence>
+  );
+};
+
+export default SplashScreen;
